refactor(Main): add StopwatchList
